% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprior.R
\name{find_dirichlet_prior}
\alias{find_dirichlet_prior}
\title{Find a dirichlet prior for box constraints}
\usage{
find_dirichlet_prior(lower, upper, ..., size_box = c(0, 10 * n), x0,
  alpha = 1, opts = OPTS, local_opts = LOCAL_OPTS, n = 1)
}
\arguments{
\item{lower}{lower bounds on the posterior proportion}

\item{upper}{upper bounds on the posterior proportion}

\item{size_box}{bounds of psuedo-sample-size}

\item{x0}{initial values of params}

\item{alpha}{tuning paramter}

\item{opts}{optimization params for nloptr}

\item{local_opts}{local optimization params for nloptr}

\item{n}{scaling parameter, sample size}
}
\description{
Find a dirichlet prior for box constraints
}
